requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
alabaster==0.7.12
astroid==1.6.6            # via pylint, pylint-celery
atomicwrites==1.3.0       # via pytest
attrs==19.3.0
babel==2.7.0
backports.functools-lru-cache==1.6.1 ; python_version == "2.7"  # via astroid, isort, pylint
certifi==2019.11.28
chardet==3.0.4
click-log==0.3.2          # via edx-lint
click==7.0                # via click-log, edx-lint
configparser==4.0.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
ddt==1.2.2
django-waffle==0.18.0
djangorestframework-jwt==1.11.0
djangorestframework==3.9.4
docutils==0.15.2
drf-yasg==1.17.0
edx-django-utils==2.0.2
edx-lint==1.4.1
edx-opaque-keys==2.0.1
enum34==1.1.6             # via astroid
factory-boy==2.12.0
faker==2.0.5              # via factory-boy
funcsigs==1.0.2           # via mock, pytest
functools32==3.2.3.post2 ; python_version == "2.7"
future==0.18.2
futures==3.1.1 ; python_version == "2.7"
httpretty==0.9.7
idna==2.8
imagesize==1.1.0
importlib-metadata==1.1.0
inflection==0.3.1
ipaddress==1.0.23         # via faker
isort==4.3.21
itypes==1.1.0
jinja2==2.10.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3  # via astroid
m2r==0.2.1
markupsafe==1.1.1
mccabe==0.6.1             # via pylint
mistune==0.8.4
mock==1.3.0
more-itertools==5.0.0
newrelic==5.4.0.132
packaging==19.2
pathlib2==2.3.5
pbr==5.4.4
pluggy==0.13.1
psutil==1.2.1
py==1.8.0
pycodestyle==2.5.0
pycryptodomex==3.9.4
pygments==2.5.2
pyjwkest==1.3.2
pyjwt==1.7.1
pylint-celery==0.3        # via edx-lint
pylint-django==0.11.1     # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==1.9.5             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pymongo==3.9.0
pyparsing==2.4.5
pyrsistent==0.15.6
pytest-cov==2.8.1
pytest-django==3.7.0
pytest==4.6.6             # via pytest-cov, pytest-django
python-dateutil==2.8.1
pytz==2019.3
pyyaml==5.2
requests==2.22.0
rest-condition==1.0.3
ruamel.ordereddict==0.4.14 ; python_version == "2.7"
ruamel.yaml.clib==0.2.0
ruamel.yaml==0.16.5
scandir==1.10.0
semantic-version==2.8.3
singledispatch==3.4.0.3   # via astroid, pylint
six==1.13.0
snowballstemmer==2.0.0
sphinx==1.8.5
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-openapi[markdown]==0.5.0
sphinxcontrib-websupport==1.1.2
stevedore==1.31.0
text-unidecode==1.3       # via faker
tox==2.9.1
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==16.7.8
wcwidth==0.1.7            # via pytest
wrapt==1.11.2             # via astroid
zipp==0.6.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
